BotnetChat messages, are no longer being .Net string encoded.. Due to .Net's fail string encoder the server is now straight up sending the messages via byte array. Encoder is still active on account names, and passwords currently.
Upgraded almost all of the integers to unsigned aswell while i was at it.

CMSG_CHAT_DROP fixed aswell, but still not active at the moment.

+++ BotNet/Botnet/Common/Common.vb
@@ -49,7 +49,7 @@ Public Module Common
         sr = Nothing
         Return outBuf
     End Function
-    Public Function GetAccountFlags(ByVal strPath As String) As Long
+    Public Function GetAccountFlags(ByVal strPath As String) As UInt32
         Dim sr As StreamReader = New StreamReader(strPath & "\" & FileAccounts)
         Dim outBuf As String = sr.ReadToEnd()
         sr.Close()
